Use import/export for GGRC.navigate
