+++ src/ggrc-client/js/components/issue/issue-unmap-item.js
@@ -11,7 +11,10 @@ import {
   batchRequests,
 } from '../../plugins/utils/query-api-utils';
 import {allowedToMap} from '../../plugins/ggrc_utils';
-import {getPageInstance} from '../../plugins/utils/current-page-utils';
+import {
+  getPageInstance,
+  navigate,
+} from '../../plugins/utils/current-page-utils';
 
 export default can.Component.extend({
   tag: 'issue-unmap-item',
@@ -122,7 +125,7 @@ export default can.Component.extend({
         );
         await relationship.unmap(true);
         if (currentObject === this.attr('issueInstance')) {
-          GGRC.navigate(this.attr('issueInstance.viewLink'));
+          navigate(this.attr('issueInstance.viewLink'));
         } else {
           this.attr('modalState.open', false);
         }
